# stario-platform/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Stario API
  description: |
    Stario / TabrikStar - AI Emotion Platform API

    Create personalized AI video greetings, face quizzes, posters, and more
    with your favorite artists.

    ## Authentication
    Most endpoints require Bearer token authentication.
    Get your token via `/auth/login` or `/auth/telegram`.

    ## Rate Limiting
    - 60 requests per minute per user
    - 10,000 requests per day per user

    ## Compliance (Uzbekistan 2025-2030)
    - Data residency: All user data stored within Uzbekistan
    - PII minimization: Face Quiz photos deleted within 3 seconds
    - Audit logs retained for 90 days

  version: 1.0.0
  contact:
    name: Stario Support
    email: support@stario.uz
servers:
  - url: https://api.stario.uz
    description: Production
  - url: https://api-staging.stario.uz
    description: Staging
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /artists:
    get:
      tags: [Artists]
      summary: List all artists
      operationId: listArtists
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
            default: UZ
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
  /artists/{artistId}:
    get:
      tags: [Artists]
      summary: Get artist details
      operationId: getArtist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Artist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '404':
          description: Artist not found
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Artist:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        stage_name:
          type: string
        bio:
          type: string
        category:
          type: string
        country:
          type: string
        avatar_url:
          type: string
        verification_status:
          type: string
          enum: [pending, approved, rejected]
        is_active:
          type: boolean
        total_videos:
          type: integer
        rating:
          type: number
...